# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ReMarket API
  version: 1.0.0
  description: |
    REST contract for the ReMarket marketplace. Monetary values are decimal
    strings in VND and timestamps are ISO-8601 UTC strings. Mutating order and
    product operations use optimistic versions. Checkout requires an
    Idempotency-Key header and replays the original result for the same payload.
servers:
  - url: /api/v1
...
paths:
...
  /support/tickets:
    get: { tags: [Trust], operationId: listSupportTickets, security: [{ bearerAuth: [] }], parameters: [{ name: status, in: query, schema: { type: string, enum: [OPEN, IN_PROGRESS, RESOLVED, CLOSED] } }, { name: type, in: query, schema: { type: string, enum: [ACCOUNT, ORDER_PROBLEM, PRODUCT, OTHER] } }, { $ref: '#/components/parameters/Page' }, { $ref: '#/components/parameters/PageSize' }], responses: { '200': { $ref: '#/components/responses/SupportTicketListSuccess' } } }
    post: { tags: [Trust], operationId: createSupportTicket, security: [{ bearerAuth: [] }], requestBody: { $ref: '#/components/requestBodies/SupportTicketBody' }, responses: { '201': { $ref: '#/components/responses/SupportTicketSuccess' } } }
  /support/tickets/{id}:
    get: { tags: [Trust], operationId: getSupportTicket, security: [{ bearerAuth: [] }], parameters: [{ $ref: '#/components/parameters/Id' }], responses: { '200': { $ref: '#/components/responses/SupportTicketSuccess' } } }
  /support/tickets/{id}/messages:
    post: { tags: [Trust], operationId: replySupportTicket, security: [{ bearerAuth: [] }], parameters: [{ $ref: '#/components/parameters/Id' }], requestBody: { $ref: '#/components/requestBodies/SupportReplyBody' }, responses: { '200': { $ref: '#/components/responses/SupportTicketSuccess' } } }
  /support/tickets/{id}/close:
    post: { tags: [Trust], operationId: closeSupportTicket, security: [{ bearerAuth: [] }], parameters: [{ $ref: '#/components/parameters/Id' }], responses: { '200': { $ref: '#/components/responses/SupportTicketSuccess' }, '409': { $ref: '#/components/responses/Error' } } }
...
  /admin/support-tickets:
    get: { tags: [Admin], operationId: adminTickets, security: [{ bearerAuth: [] }], parameters: [{ name: status, in: query, schema: { type: string, enum: [OPEN, IN_PROGRESS, RESOLVED, CLOSED, ALL] } }, { name: type, in: query, schema: { type: string, enum: [ACCOUNT, ORDER_PROBLEM, PRODUCT, OTHER, ALL] } }, { name: assigned_admin_id, in: query, schema: { type: string, description: Admin UUID or UNASSIGNED } }, { name: from, in: query, schema: { type: string, format: date } }, { name: to, in: query, schema: { type: string, format: date } }, { $ref: '#/components/parameters/Page' }, { $ref: '#/components/parameters/PageSize' }], responses: { '200': { $ref: '#/components/responses/AdminSupportTicketListSuccess' } } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
    Id: { name: id, in: path, required: true, schema: { type: string, format: uuid } }
...
    Page: { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
    PageSize: { name: page_size, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
  requestBodies:
...
    SupportTicketBody:
      required: true
      content: { application/json: { schema: { $ref: '#/components/schemas/SupportTicketInput' } } }
    SupportReplyBody:
      required: true
      content: { application/json: { schema: { $ref: '#/components/schemas/SupportReplyInput' } } }
...
  responses:
...
    SupportTicketListSuccess:
      description: Paginated support tickets
      content: { application/json: { schema: { allOf: [{ $ref: '#/components/schemas/SuccessEnvelope' }, { type: object, properties: { data: { $ref: '#/components/schemas/SupportTicketListData' } } }] } } }
    SupportTicketSuccess:
      description: Support ticket detail
      content: { application/json: { schema: { allOf: [{ $ref: '#/components/schemas/SuccessEnvelope' }, { type: object, properties: { data: { $ref: '#/components/schemas/SupportTicketDetail' } } }] } } }
...
    AdminSupportTicketListSuccess:
      description: Paginated admin support tickets
      content: { application/json: { schema: { allOf: [{ $ref: '#/components/schemas/SuccessEnvelope' }, { type: object, properties: { data: { $ref: '#/components/schemas/AdminSupportTicketListData' } } }] } } }
...
    Error:
      description: Business or validation error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  schemas:
...
    SupportTicketInput:
      type: object
      required: [type, subject, message]
      additionalProperties: false
      properties:
        type: { type: string, enum: [ACCOUNT, ORDER_PROBLEM, PRODUCT, OTHER] }
        subject: { type: string, minLength: 5, maxLength: 150, example: Cần hỗ trợ đơn hàng }
        message: { type: string, minLength: 1, maxLength: 5000 }
        order_id: { type: string, format: uuid }
    SupportReplyInput:
      type: object
      required: [message]
      additionalProperties: false
      properties: { message: { type: string, minLength: 1, maxLength: 5000 } }
...
    IsoDateTime: { type: string, format: date-time }
...
    TicketType: { type: string, enum: [ACCOUNT, ORDER_PROBLEM, PRODUCT, OTHER] }
    TicketStatus: { type: string, enum: [OPEN, IN_PROGRESS, RESOLVED, CLOSED] }
    PageMeta:
      type: object
      required: [page, page_size, total, total_pages]
      additionalProperties: false
      properties:
        page: { type: integer, minimum: 1 }
        page_size: { type: integer, minimum: 1, maximum: 100 }
        total: { type: integer, minimum: 0 }
        total_pages: { type: integer, minimum: 1 }
...
    SupportTicketListItem:
      type: object
      required: [id, code, subject, type, status, order_id, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        code: { type: string }
        subject: { type: string }
        type: { $ref: '#/components/schemas/TicketType' }
        status: { $ref: '#/components/schemas/TicketStatus' }
        order_id: { type: string, format: uuid, nullable: true }
        created_at: { $ref: '#/components/schemas/IsoDateTime' }
        updated_at: { $ref: '#/components/schemas/IsoDateTime' }
    SupportMessageSender:
      type: object
      required: [id, name, role]
      additionalProperties: false
      properties:
        id: { type: string, description: UUID for admins and users; the literal `support` when admin identity is hidden. }
        name: { type: string }
        role: { type: string, enum: [USER, ADMIN] }
    SupportMessage:
      type: object
      required: [id, sender, message, created_at]
      additionalProperties: false
      properties:
        id: { type: string, format: uuid }
        sender: { $ref: '#/components/schemas/SupportMessageSender' }
        message: { type: string }
        created_at: { $ref: '#/components/schemas/IsoDateTime' }
    SupportTicketDetail:
      allOf:
        - { $ref: '#/components/schemas/SupportTicketListItem' }
        - type: object
          required: [resolution_note, resolved_at, closed_at, assigned_admin_name, messages, allowed_actions]
          properties:
            resolution_note: { type: string, nullable: true }
            resolved_at: { allOf: [{ $ref: '#/components/schemas/IsoDateTime' }], nullable: true }
            closed_at: { allOf: [{ $ref: '#/components/schemas/IsoDateTime' }], nullable: true }
            assigned_admin_name: { type: string, nullable: true }
            messages: { type: array, items: { $ref: '#/components/schemas/SupportMessage' } }
            allowed_actions: { type: array, items: { type: string, enum: [reply, close, reopen, resolve, progress] } }
...
    AdminSupportTicketItem:
      allOf:
        - { $ref: '#/components/schemas/SupportTicketListItem' }
        - type: object
          required: [user, assigned_admin]
          properties:
            user: { type: object, required: [id, name], additionalProperties: false, properties: { id: { type: string, format: uuid }, name: { type: string } } }
            assigned_admin: { type: object, nullable: true, required: [id, name], additionalProperties: false, properties: { id: { type: string, format: uuid }, name: { type: string } } }
...
    SupportTicketListData:
      type: object
      required: [items, meta]
      additionalProperties: false
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/SupportTicketListItem' } }
        meta: { $ref: '#/components/schemas/PageMeta' }
...
    AdminSupportTicketListData:
      type: object
      required: [items, meta]
      additionalProperties: false
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/AdminSupportTicketItem' } }
        meta: { $ref: '#/components/schemas/PageMeta' }
...
    Meta:
      type: object
      required: [request_id]
      properties:
        request_id: { type: string, format: uuid }
        page: { type: integer, minimum: 1 }
        page_size: { type: integer, minimum: 1, maximum: 100 }
        total: { type: integer, minimum: 0 }
        total_pages: { type: integer, minimum: 1 }
    SuccessEnvelope:
      type: object
      required: [success, data, meta]
      properties:
        success: { type: boolean, enum: [true] }
        data: {}
        meta: { $ref: '#/components/schemas/Meta' }
    ApiErrorCode:
      type: string
      description: Stable machine-readable error code shared by the API and web client.
      enum:
        - VALIDATION_ERROR
        - UNAUTHORIZED
        - FORBIDDEN
        - EMAIL_NOT_VERIFIED
        - ACCOUNT_LOCKED
        - NOT_FOUND
        - PRODUCT_NOT_AVAILABLE
        - PRICE_CHANGED
        - INVALID_ORDER_TRANSITION
        - ORDER_EXPIRED
        - VERSION_CONFLICT
        - REVIEW_NOT_ALLOWED
        - IDEMPOTENCY_CONFLICT
        - RETRY_LATER
        - RATE_LIMITED
        - SESSION_EXPIRED
        - INTERNAL
      example: VERSION_CONFLICT
    ErrorEnvelope:
      type: object
      required: [success, error, meta]
      properties:
        success: { type: boolean, enum: [false] }
        error:
          type: object
          required: [code, message]
          properties:
            code: { $ref: '#/components/schemas/ApiErrorCode' }
            message: { type: string, example: Dữ liệu đã thay đổi, vui lòng tải lại. }
            details: { type: object, additionalProperties: true }
        meta: { $ref: '#/components/schemas/Meta' }